feature: code Circuit.as_unitary() - implement with braket.circuits.unitary_calculation.calculate_unitary()
