+++ src/braket/circuits/circuit.py
@@ -35,6 +35,7 @@ from braket.circuits.observables import TensorProduct
 from braket.circuits.qubit import QubitInput
 from braket.circuits.qubit_set import QubitSet, QubitSetInput
 from braket.circuits.result_type import ObservableResultType, ResultType
+from braket.circuits.unitary_calculation import calculate_unitary
 from braket.ir.jaqcd import Program
 
 SubroutineReturn = TypeVar(
@@ -876,6 +877,40 @@ the number of qubits in target_qubits must be the same as defined by the multi-q
             results=ir_results,
             basis_rotation_instructions=ir_basis_rotation_instructions,
         )
+
+    def as_unitary(self) -> np.ndarray:
+        """
+        Returns the unitary matrix representation of the entire circuit.
+        *Note*: The performance of this method degrades with qubit count. It might be slow for
+        qubit count > 10.
+
+        Returns:
+            np.ndarray: A numpy array with shape (2^qubit_count, 2^qubit_count) representing the
+                circuit as a unitary. *Note*: For an empty circuit, an empty numpy array is
+                returned (`array([], dtype=complex128)`)
+
+        Raises:
+            TypeError: If circuit is not composed only of `Gate` instances,
+                i.e. a circuit with `Noise` operators will raise this error.
+
+        Examples:
+            >>> circ = Circuit().h(0).cnot(0, 1)
+            >>> circ.as_unitary()
+            array([[ 0.70710678+0.j,  0.70710678+0.j,  0.        +0.j,
+                     0.        +0.j],
+                   [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
+                    -0.70710678+0.j],
+                   [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
+                     0.70710678+0.j],
+                   [ 0.70710678+0.j, -0.70710678+0.j,  0.        +0.j,
+                     0.        +0.j]])
+        """
+        qubits = self.qubits
+        if not qubits:
+            return np.zeros(0, dtype=complex)
+        qubit_count = max(qubits) + 1
+
+        return calculate_unitary(qubit_count, self.instructions)
 
     def _copy(self) -> Circuit:
         copy = Circuit().add(self.instructions)
