+++ src/braket/circuits/unitary_calculation.py
@@ -11,10 +11,12 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
+from typing import Iterable
+
 import numpy as np
 
-from braket.circuits.circuit import Circuit
 from braket.circuits.gate import Gate
+from braket.circuits.instruction import Instruction
 from braket.circuits.qubit_set import QubitSet
 
 
@@ -38,45 +40,30 @@ def _einsum_subscripts(targets: QubitSet, qubit_count: int) -> str:
     )
 
 
-def as_unitary(circuit: Circuit) -> np.ndarray:
+def calculate_unitary(qubit_count: int, instructions: Iterable[Instruction]) -> np.ndarray:
     """
-    Returns the unitary matrix representation of the entire `circuit`.
+    Returns the unitary matrix representation for all the `instructions` with a given
+    `qubit_count`.
     *Note*: The performance of this method degrades with qubit count. It might be slow for
     qubit count > 10.
 
     Args:
-        circuit (Circuit): The circuit from which to get the unitary matrix.
+        qubit_count (int): Total number of qubits, enough for all the `instructions`.
+        instructions (Iterable[Instruction]): The instructions for which the unitary matrix
+            will be calculated.
 
     Returns:
         np.ndarray: A numpy array with shape (2^qubit_count, 2^qubit_count) representing the
-            `circuit` as a unitary. *Note*: For an empty `circuit`, an empty numpy array is
-            returned (`array([], dtype=complex128)`).
+            `instructions` as a unitary.
 
     Raises:
-        TypeError: If `circuit` is not composed only of `Gate` instances,
+        TypeError: If `instructions` is not composed only of `Gate` instances,
             i.e. a circuit with `Noise` operators will raise this error.
-
-    Examples:
-        >>> circ = Circuit().h(0).cnot(0, 1)
-        >>> as_unitary(circ)
-        array([[ 0.70710678+0.j,  0.70710678+0.j,  0.        +0.j,
-                 0.        +0.j],
-               [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
-                -0.70710678+0.j],
-               [ 0.        +0.j,  0.        +0.j,  0.70710678+0.j,
-                 0.70710678+0.j],
-               [ 0.70710678+0.j, -0.70710678+0.j,  0.        +0.j,
-                 0.        +0.j]])
     """
-    qubits = circuit.qubits
-    if not qubits:
-        return np.zeros(0, dtype=complex)
-    qubit_count = max(qubits) + 1
-
     unitary = np.eye(2 ** qubit_count, dtype=complex)
     un_tensor = np.reshape(unitary, qubit_count * [2, 2])
 
-    for instr in circuit.instructions:
+    for instr in instructions:
         if not isinstance(instr.operator, Gate):
             raise TypeError("Only Gate operators are supported to build the unitary")
